+++ reducers/spaces.js
@@ -3,10 +3,7 @@ import { CREATE_GROUP_WEB3_SUCCESS, SPACE_LOAD_SUCCESS, SPACE_LOAD_POSTS_SUCCESS
 const initialState = {
     data: {
         '0x68cCe7ad9beDf41BEE8A7D1b758EDBb8AEeFb1b9': new Space('0x68cCe7ad9beDf41BEE8A7D1b758EDBb8AEeFb1b9', 'The Lobby', null, 3)
-    },
-    profiles: {},
-    createdSpaces: [
-    ],
+    }
 }
 
 function Space(addr, name, thread, chainId) {
@@ -60,17 +57,6 @@ export default function reduce(state = initialState, action) {
                         ...space,
                         posts,
                     }
-                }
-            }
-        }
-        case 'ADD_USER_PROFILE': {
-            const { user, profile } = action.payload
-            const { profiles } = state
-            return {
-                ...state,
-                profiles: {
-                    ...profiles,
-                    [user]: profile
                 }
             }
         }
